requirements.txt
absl-py==0.10.0
aiohttp==3.6.3
alembic==1.4.3
annoy==1.17.0
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cryptography==3.1.1
docker-pycreds==0.4.0
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
gast==0.3.3
gitdb==4.0.5
gitpython==3.1.11
google-auth-oauthlib==0.4.1
google-auth==1.21.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imdbpy==2020.9.25
importlib-metadata==2.0.0
isort==5.5.3
itsdangerous==1.1.0
jinja2==2.11.2
keras-preprocessing==1.1.2
lazy-object-proxy==1.4.3
lxml==4.5.2
mako==1.1.3
markdown==3.2.2
markupsafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
multipledispatch==0.6.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pathtools==0.1.2
pip==20.2.4
promise==2.3
protobuf==3.13.0
psutil==5.7.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyjwt==1.7.1
pylint==2.6.0
pyopenssl==19.1.0
pysocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests==2.24.0
rsa==4.6
scipy==1.5.2
sentry-sdk==0.19.1
setuptools==50.3.1.post20201107
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
sqlalchemy==1.3.19
subprocess32==3.5.4
tensorboard-plugin-wit==1.7.0
tensorboard==2.3.0
tensorflow-estimator==2.3.0
tensorflow-recommenders==0.2.0
tensorflow==2.3.1
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.10
wandb==0.10.8
watchdog==0.10.3
werkzeug==1.0.1
wheel==0.35.1
wrapt==1.12.1
wtforms==2.3.3
yarl==1.6.2
zipp==3.4.0
